Move jsonlog to a subpackage of jsonfilelog
